pyproject.toml
[project]
name = "fakedin"
version = "0.1.0"
description = "Utilities to create fake resumes and job descriptions"
authors = [
    {name = "B.T. Franklin", email = "brandon.franklin@gmail.com"},
]
dependencies = [
    "faker>=37.0.0",
    "fpdf2>=2.8.2",
    "jinja2>=3.1.6",
    "markdown>=3.7",
    "openai>=1.65.5",
    "promptdown>=0.9.0",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
